requirements.txt
sounddevice
JACK-Client
numpy
jack-server